requirements.txt
flake8
j2cli[yaml]
yamllint
pyyaml